chore: modify utility class to to deconstruct nested ifs

+++ src/main/java/employee/service/EmployeeService.java
@@ -2,6 +2,6 @@ package employee.service;
 
 public interface EmployeeService {
 
-    public Object getEmployee(String id, int version);
+    public Object getEmployee(String id, double version);
 
 }
